Move browser testing versions to it's own file, and programatically create grunt config for saucelabs-mocha
